HCF-1167 make lint/format should fail on error
When the program (golint / goimports) reports an error — for example
because they were not installed - we should correctly fail.

This fixes the issue by getting rid of the pipe so `errexit` is still in
effect.  While we're there, get rid of the `wc` too, since we can just
test for an empty file.  That means always removing the file during the
trap, rather than early, which lets us unconditionally cat it too.

Also fixes problems identified by shellcheck.