Rename github.com/suse to github.com/SUSE
This is the canonical spelling of our Github org name, and we need to use
it consistently in case repos get synced into a GOPATH on a non-case-sensitive
filesystem; otherwise Go will complain.

This change has been made by a global search-and-replace in an editor and
not via glide, which wanted to make much more intensive changes. Those can
wait until we switch to dep instead.